Moved Poll to a models directory. Restored saving of new polls to the database. Second poll creation is successfully blocked.
