+++ polls_repository.py
@@ -8,6 +8,7 @@
 import sqlite3
 from papertrail import logger
 from db_constants import *
+from models.poll import Poll
 
 def get_connection(db_path):
     return sqlite3.connect(db_path)
@@ -28,12 +29,12 @@ class PollsRepository(object):
         self.conn.execute(SQL_CREATE_VOTES_TABLE)
         self.conn.commit()
 
-    def create_poll(self, poll):
+    def create_poll(self, db_conn, poll):
         try:
-            self.conn.execute(SQL_CREATE_POLL % (poll.id, poll.team_id, poll.channel_id, \
+            db_conn.execute(SQL_CREATE_POLL % (poll.poll_id, poll.team_id, poll.channel_id, \
             poll.name, poll.is_open, poll.owner_user_id))
-            self.conn.commit()
-            logger.info("Successfully inserted poll " + poll.id + " into the database.")
+            db_conn.commit()
+            logger.info("Successfully inserted poll " + poll.poll_id + " into the database.")
         except sqlite3.Error as e:
             logger.info("Error inserting poll into the database. " + str(e))
 
@@ -46,7 +47,7 @@ class PollsRepository(object):
                 poll = Poll(row[0], row[1], row[2], row[3], row[4], row[5])
 
 
-            logger.info("Successfully retrieved poll " + poll.id + " from the database.")
+            logger.info("Successfully retrieved poll " + poll.poll_id + " from the database.")
 
             return poll
         except sqlite3.Error as e:
@@ -72,13 +73,3 @@ class PollsRepository(object):
         except sqlite3.Error as e:
             logger.info("SQL Error " + str(e))
             return None
-
-class Poll(object):
-    def __init__(self, poll_id, team_id, channel_id, name, is_open, owner_user_id):
-        self.id = poll_id
-        self.team_id = team_id
-        self.channel_id = channel_id
-        self.name = name
-        self.is_open = is_open
-        self.owner_user_id = owner_user_id
-
